medix-test-branch-b-ver-0.9
Added Nurses and their dashboard, they can access appointments, list, medical records, and the first 5 available nurses in the list can be seen.

Also patch the bug in the working days not showing the start_time & close_time when toggled.

Also patched the bug with the working days not correctly showing whether the start_time / close time is AM or PM. (in doctors/page.tsx)

Fixed an admin action wherein when adding a new staff member it always gives them the role of a doctor. As of now. the only viable role if your a staff member is a nurse.

Lab technician is cancelled or in progress for now cuz we need to make the lab dashboard if we want that.
Physical therapist might be a mandatory so i'll research on what data or being taken or stored when your a physical therapist.

I also plant to add a Patient to Staff Ratio charts  in the future