Patch 7 - Fixed some bugs in the code of this files, below.
Modified the record/billing/page.tsx file - changed some of the code, it has an error on duplicate keys in the tb table.

General.ts - Modified the deletebyID code to do a recalculation whenever the actiondialog delete is called.

Navbar - modified the code to do multiple checks and a delay function to wait for the server to render first and push towards the client side. The hydration error is caused by the lack of loading in the system. Maybe i'll add a loading functionality in the future.

BillsContainer - changed the code in order for the values to always reset to zero when a bill is deleted.

Generate final bills - the main problem was the event call function modified it to either do a callback/refresh whenever a bill is added. The error of the button not working seems to be that the initial code doesn't check the state of the table when a new bill is added thats why you need to refresh before.