medix-branch-3-ver-0.9
Improved the appointments table allowing users to see the priority of the there appointments

Improved the create appointments code to utilize djiktra algorithm to decide the priority level of appointments, via keyword matching and assigning them automatically to the correct department

Added a referral management system, allowing doctors to create there referrals in case a patient needed one.

I plan to integrate FSM in the referral management in the future.

But for now The system has a working djikstra algorithm already implemented.

Bug to fix in the future when creating appointments, in the view table fix the Date Column. It has correct dates but the time is kinda bugging me.

Improved the looks of the sidebar, to make it more sleek and interactive.

Fixed some bugs in some parts of the system.

Next Update: Swiftly integrate paymongo to the system and create the Invoice Sidebar component later.