Patch 9  - Payment
Integrated paymongo credit/card payment method.

Gcash & Other payment method's are unavailable as it needs verification or DTI ID to make the paymongo account verified.

Addded a billing details page when viewing the bills.

When paying payments the status of the payment is not yet updated via the database. Maybe in a later update we'll do that.

Use only test cards / test transactions to test the payment functions.

Codebase is not yet clean, maybe in a later patch.