Patch 11 - Allow Nurses to Book Appointments
Update:
Created a new nurse-appointment-wrapper.tsx and nurse-book-appointment. In order to allow nurses to assist patient's on booking appointments. They can do this in the patient's section of the sidebar.

Improved the design of the view-appointment component and added more info such as doctor information, self-assist or assisted appointment, workday time, etc.