medix-test-branch-3-ver.0.11
Added Made the Appointment Analyzer More Dynamic (but it's still buggy)
Relocated the Advanced Settings of the Analyzer to the Admin's Settings Page.
Added A custom Keyword Manager to let admin's for now to create groups and in those groups register keywords that the appointment analyzer will use.
Changed the Font to FiraCode
Added a auto-logout if there are no user activity in a specified time (minutes). The time is different on the user role.