medix-test-branch-ver.0.13
Changed alot of major codes on the priority analyzer, made the override tag to be more accurate, but it's still buggy

Removed the draggable component on both booking appointment codes.
Changed both of the design to be more simpler with green patient & blue for nurse.

Changed calendar from default html to Shad CN Calendar

Made the shad cn calendar dynamic and the time slots too are also dynamic now.

So that means based on the data of the doctor's working days, it would auto cross out the days on the calendar in which the doctor is unavailable, the time is also dynamic so when a doctor is only available for example for halfday, the day on the calendar auto closes...

The dynamic nature of the calendar ranges thru 3 months... passed the 3 months all dates are disabled.